Delete Deck.java and Hand.java Update Tests and imports
